refactor: error 패키지 구조 수정
error 패키지는 HTTP 관련 예외를 의미하므로 의도를 명확히 하기 위해 http 패키지 하위로 이동함
