+++ src/main/java/codesquad/http/error/HttpRequestParseException.java
@@ -1,4 +1,4 @@
-package codesquad.error;
+package codesquad.http.error;
 
 public class HttpRequestParseException extends RuntimeException {
