+++ src/main/java/codesquad/http/error/QueryParseException.java
@@ -1,4 +1,4 @@
-package codesquad.error;
+package codesquad.http.error;
 
 public class QueryParseException extends RuntimeException {
