+++ src/main/java/codesquad/http/error/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package codesquad.error;
+package codesquad.http.error;
 
 public class ResourceNotFoundException extends RuntimeException {
