+++ src/main/java/codesquad/http/error/UnSupportedMediaTypeException.java
@@ -1,4 +1,4 @@
-package codesquad.error;
+package codesquad.http.error;
 
 public class UnSupportedMediaTypeException extends RuntimeException {
