+++ src/main/java/codesquad/http/parser/HttpParser.java
@@ -3,7 +3,7 @@ package codesquad.http.parser;
 import static codesquad.http.HttpConstraints.HEADER_DELIMITER;
 import static codesquad.utils.StringUtils.BLANK;
 
-import codesquad.error.HttpRequestParseException;
+import codesquad.http.error.HttpRequestParseException;
 import codesquad.http.message.HttpRequest;
 import codesquad.http.property.HttpMethod;
 import codesquad.http.property.HttpVersion;
